+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: kjurkows <kjurkows@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/15 20:39:46 by kjurkows          #+#    #+#             */
-/*   Updated: 2026/06/16 14:22:29 by kjurkows         ###   ########.fr       */
+/*   Updated: 2026/06/16 17:56:21 by kjurkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long unsigned int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	long unsigned int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len])
